+++ build.rs
@@ -161,28 +161,19 @@ fn compile_simd_c() {
     let mut build = cc::Build::new();
     build.opt_level(3);
 
-    // `-march=native` is not supported on Apple M1
-    #[cfg(not(all(target_os = "macos", target_arch = "aarch64")))]
-    {
-        fn guess_target_cpu() -> String {
-            // Copied and adapted from https://github.com/alexcrichton/proc-macro2/blob/4173a21dc497c67326095e438ff989cc63cd9279/build.rs#L115
-            // Licensed under Apache-2.0 + MIT (compatible because we're MIT)
-            let rustflags = env::var_os("RUSTFLAGS").or(env::var_os("CARGO_ENCODED_RUSTFLAGS"));
-            if let Some(rustflags) = rustflags {
-                for mut flag in rustflags.to_string_lossy().split(&[' ', '\u{1f}'][..]) {
-                    if flag.starts_with("-C") {
-                        flag = &flag["-C".len()..];
-                    }
-                    if flag.starts_with("target-cpu=") {
-                        return flag["target-cpu=".len()..].to_owned();
-                    }
-                }
-            }
-
-            "native".to_string()
+    match env::var("RUST_REED_SOLOMON_ERASURE_ARCH") {
+        Ok(arch) => {
+            // Use explicitly specified environment variable as architecture.
+            build.flag(&format!("-march={}", arch));
         }
-        build.flag(&format!("-march={}", guess_target_cpu()));
+        Err(_error) => {
+            // On x86-64 enabling Haswell architecture unlocks useful instructions and improves performance
+            // dramatically while allowing it to run ony modern CPU.
+            #[cfg(target_arch = "x86_64")]
+            build.flag(&"-march=haswell");
+        }
     }
+
     build
         .flag("-std=c11")
         .file("simd_c/reedsolomon.c")