+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 5.0.2
+* Add support for `RUST_REED_SOLOMON_ERASURE_ARCH` environment variable and stop using `native` architecture for SIMD code
+  - See [PR #98](https://github.com/rust-rse/reed-solomon-erasure/pull/98)
+
 ## 5.0.1
 - The `simd-accel` feature now builds on M1 Macs
   - See [PR #92](https://github.com/rust-rse/reed-solomon-erasure/pull/92)